Evaluate the expected number of trees in EvalOneIter
EvalOneIter takes an argument iter which, per the related header
file, indicates the iteration number to evaluate. Prior to this patch
the only effect this argument had was to change the number printed
in the output. Update the function to evaluate the requested boosting
iteration.

This has no effect on existing training as the desired iteration to
evaluate is always passed so the output looks correct to the user.