Extract interaction constraints, enable it for approx.
* Extract interaction constraints from split evaluator.

The primary reason for doing so is that it copies the `num_feature` parameter,
which makes serialization and parameter validation difficult.  Also, as it
should be used for selecting feature, like column sampler, instead of computing
weight.

* clean up for colmaker.

Remove support for `parallel_option` and `cache_opt`.  Now we use whatever
settings that are default before this PR.  As these parameters are never
documented nor actually maintained.

* Enable for approx.