Add Optuna badge to README.md
Hi!

Optuna (www.optuna.org) is a new hyperparameter optimization library, and we've written an integration module for XGBoost to make it easy to use Optuna to search for good hyperparameter settings and prune unpromising trials. We're looking for ways to let users of XGboost know this is available and thought a badge to show Optuna integration is available would be helpful.

If there are other ways you would recommend reaching out to XGBoost users to let them know about Optuna, please let us know.

Thanks!