Simplify CSRAdapter
Whilst working on #5854, I noticed that the nelem argument supplied to
`XGDMatrixCreateFromCSREx` seems to have no effect whatsoever.

`CSRAdapterBatch` currently doesn't seem to store this field, let alone
do anything with it.

As such, it seems like it can be removed entirely.

This PR removes references to it from CSRAdapter and CSRAdapterBatch.
The signature of the C API functions has not been modified to ensure
backwards compatibility.