Revert OMP guard.
The guard protects the global variable from being changed by XGBoost.  But this lead to a
bug that the `n_threads` parameter is no longer used after the first iteration.  This is
due to the fact that `omp_set_num_threads` is only called once in `Learner::Configure` at
the beginning of the training process.

The guard is still useful for gpu_id, since this is called all the times in our code base
doesn't matter which iteration we are currently running.