Fix bug when num_row is zero in CSRAdapter
I noticed a difference in the behavior of `XGDMatrixCreateFromCSREx` and `XGDMatrixCreateFromCSCEx` when I was using c_api.

`CSRAdapter` does not generate an appropriate matrix when `num_row` is zero.  This is not what the documentation says.
This creates an empty Matrix and returns, raise error in learning `Booster`.

```
...out/xgboost/src/learner.cc:602: Check failed: mparam_.num_feature != 0 (0 vs. 0) : 0 feature is supplied.  Are you using raw Booster interface?
```

doc: https://xgboost.readthedocs.io/en/latest/dev/c__api_8h.html#a91d3ab5fa109f418a368d66ad2b99d3c

> num_col	number of columns; when it's set to kAdapterUnknownSize, then guess from data

This PR is a modification to make the `CSRAdapter` behave the same as the `CSCAdapter`.