Use context in `SetInfo`.
* Use the name `Context`.
* Pass a context object into `SetInfo`.
* Add context to proxy matrix.
* Add context to iterative DMatrix.

This is to remove the use of default number of threads during `SetInfo` as a follow up on
removing global omp variable, while prepare for CUDA stream semantic.  Currently XGBoost
uses legacy CUDA stream along with a device sync in `gpu_hist`, we will gradually remove
them in the future in favor of non-blocking streams.