Define bin type.
This PR defines a `bst_bin_t` similar to `bst_feature_t` and friends. The conversions
between signedness in function calls and returns are confusing sometimes. I hope that we
can unify them in the future. Also, this provides a better document for the parameters
like `index_base`, at least we should know from the first glance that it's referring to
something related to histogram bins.