Obtain CSR matrix from DMatrix.
* Obtain CSR matrix from DMatrix.
* Obtain gradient index from Quantile DMatrix.

This is mostly for testing at higher level. Right now we rely on training a booster to
infer that the DMatrix is correctly constructed. With this PR, we can ease the testing
process. Also, this has been requested before.

The return value from Quantile DMatrix is histogram index instead of cut values as we
shift the cut value to include min_values, which is not very useful for language bindings.

Unlike most of XGBoost C functions, caller of C API is required to allocate the memory
itself instead of using thread local memory from XGBoost. This is to avoid allocating a
huge memory buffer that can not be freed until exiting the thread.

External memory is not supported.