Fix loading GPU pickle with a CPU-only xgboost distribution.
We can handle loading the pickle on a CPU only machine if the XGBoost is built with CUDA
enabled (Linux and Windows PyPI package), but not if the distribution is CPU-only (macos
PyPI package).

To reproduce the error: train a model with GPU hist. Load it with a XGBoost build that
doesn't have CUDA capability.