Initial implementation of the ordinal recoder.
No integration yet, just the recoder. Retrieving the storing the data in xgboost is more
complicated than the recoder itself, will upstream it in future PRs.

The recoder still uses some utilities in XGBoost like the `Span` class and an iterator. If
we want to extract it to a different project, we can find a different implementation of
these utilities.

Tests require a container class in XGBoost, we can merge the container into the encoder
module if needed. At the moment, the encoder is view-only and doesn't own any
memory. Larger and more sophisticated tests will be done after the Python and R
integration is finished.